LRAC-12616 Create '__tests__' folder to Attributes and move files to it
